// src/store/crm/types.tsx
export enum crmActionType {
  GET_STAGE_BEGIN = "GET_STAGE_BEGIN",
  GET_STAGE_SUCCESS = "GET_STAGE_SUCCESS",
  GET_STAGE_ERROR = "GET_STAGE_ERROR",
  CREATE_STAGE_BEGIN = "CREATE_STAGE_BEGIN",
  CREATE_STAGE_SUCCESS = "CREATE_STAGE_SUCCESS",
  CREATE_STAGE_ERROR = "CREATE_STAGE_ERROR",
  DELETE_STAGE_BEGIN = "DELETE_STAGE_BEGIN",
  DELETE_STAGE_SUCCESS = "DELETE_STAGE_SUCCESS",
  DELETE_STAGE_ERROR = "DELETE_STAGE_ERROR",
  RUN_QUALIFICATION_BEGIN = "RUN_QUALIFICATION_BEGIN",
  RUN_QUALIFICATION_SUCCESS = "RUN_QUALIFICATION_SUCCESS",
  RUN_QUALIFICATION_ERROR = "RUN_QUALIFICATION_ERROR",
}